backend/pyproject.toml
[tool.poetry]
name = "recruiterai-backend"
version = "0.1.0"
description = "A comprehensive Django-based recruitment platform backend with REST API and Swagger documentation"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "config"}, {include = "apps"}]
include = ["manage.py", "requirements.txt"]

[tool.poetry.dependencies]
python = "^3.11"
Django = "^4.2.7"
djangorestframework = "^3.14.0"
drf-spectacular = "^0.26.5"
python-decouple = "^3.8"
psycopg2-binary = "^2.9.9"
django-cors-headers = "^4.3.1"
Pillow = "^10.1.0"
gunicorn = "^21.2.0"

[tool.poetry.group.dev.dependencies]
black = "^23.0.0"
flake8 = "^6.0.0"
isort = "^5.12.0"
pytest = "^7.4.0"
pytest-django = "^4.5.0"
pre-commit = "^3.3.0"

[tool.poetry.group.test.dependencies]
pytest = "^7.4.0"
pytest-django = "^4.5.0"
pytest-cov = "^4.1.0"
factory-boy = "^3.3.0"

[tool.poetry.scripts]
start = "manage:main"
test = "pytest:main"
migrate = "django.core.management:execute_from_command_line"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["config", "apps"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
addopts = "-v --tb=short --strict-markers"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

